backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.1
django-cors-headers==2.5.3
django-filter==2.1.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
reportlab==3.5.23
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1